intelligent-vm-manager/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
typing-extensions>=4.0.0
dataclasses>=0.6 